distractors_benchmark/requirements.txt
bsuite==0.3.5
dm-acme==0.2.0
dm-env==1.4
dm-reverb==0.2.0
dm-sonnet==2.0.0
frozendict==1.2
multiprocess==0.70.11.1
numpy==1.19.5
tensorflow==2.1.0
trfl==1.1.0
wandb==0.10.20